feat: enhance RAG system with HNSW algorithm support and CLI utilities
Major improvements to the RAG (Retrieval-Augmented Generation) system:

Database & Storage Enhancements:
- Add HNSW (Hierarchical Navigable Small World) algorithm support for efficient vector similarity search
- Implement SQLiteVectorStoreHNSW with optimized HNSW graph structure
- Add new schema definitions for HNSW vector storage with Float32Array support
- Enhance database migration and configuration for HNSW parameters (M, Ml, EfSearch)

CLI Command Utilities:
- Add 'index' command for indexing plugins with configurable embedding models
- Add 'search' command for semantic search across indexed plugins
- Add 'list-collection' command to display all available vector collections
- Support for various embedding models and API configurations

API & Export Improvements:
- Simplify RAG API with cleaner, more intuitive interfaces
- Optimize plugin RAG manager with better metadata handling
- Enhance export/import functionality for vector collections
- Remove deprecated plugin_indexer.go in favor of new command structure

Algorithm & Performance:
- Implement pure Go HNSW algorithm with cosine similarity calculations
- Add distance function abstraction for different similarity metrics
- Optimize vector operations and graph traversal performance
- Add comprehensive test coverage for HNSW operations

Breaking Changes:
- Updated vector store schema to support HNSW-specific fields
- Modified collection configuration structure
- Enhanced embedding interface for better model compatibility

This update significantly improves the RAG system's search performance and provides
a complete CLI toolkit for managing vector collections and plugin indexing.

+++ common/utils/cosine_similarity_test.go
@@ -9,8 +9,8 @@ import (
 func TestCosineSimilarity(t *testing.T) {
 	// 测试用例1: 两个相似的向量 (方向接近)
 	t.Run("similar vectors", func(t *testing.T) {
-		vec1 := []float64{1.0, 2.0, 3.0, 4.0}
-		vec2 := []float64{2.0, 4.0, 6.0, 8.0} // vec2 是 vec1 的两倍
+		vec1 := []float32{1.0, 2.0, 3.0, 4.0}
+		vec2 := []float32{2.0, 4.0, 6.0, 8.0} // vec2 是 vec1 的两倍
 
 		similarity, err := CosineSimilarity(vec1, vec2)
 		if err != nil {
@@ -18,7 +18,7 @@ func TestCosineSimilarity(t *testing.T) {
 		}
 
 		// 余弦相似度应该非常接近 1.0
-		if math.Abs(similarity-1.0) > 1e-10 {
+		if math.Abs(float64(similarity-1.0)) > 1e-10 {
 			t.Errorf("Expected similarity close to 1.0, got %f", similarity)
 		}
 
@@ -27,8 +27,8 @@ func TestCosineSimilarity(t *testing.T) {
 
 	// 测试用例2: 两个正交的向量 (方向垂直)
 	t.Run("orthogonal vectors", func(t *testing.T) {
-		vec3 := []float64{1.0, 0.0}
-		vec4 := []float64{0.0, 1.0}
+		vec3 := []float32{1.0, 0.0}
+		vec4 := []float32{0.0, 1.0}
 
 		similarity, err := CosineSimilarity(vec3, vec4)
 		if err != nil {
@@ -36,7 +36,7 @@ func TestCosineSimilarity(t *testing.T) {
 		}
 
 		// 正交向量的余弦相似度应该为 0.0
-		if math.Abs(similarity) > 1e-10 {
+		if math.Abs(float64(similarity)) > 1e-10 {
 			t.Errorf("Expected similarity to be 0.0, got %f", similarity)
 		}
 
@@ -45,8 +45,8 @@ func TestCosineSimilarity(t *testing.T) {
 
 	// 测试用例3: 两个方向相反的向量
 	t.Run("opposite vectors", func(t *testing.T) {
-		vec5 := []float64{1.0, 1.0, 1.0}
-		vec6 := []float64{-1.0, -1.0, -1.0}
+		vec5 := []float32{1.0, 1.0, 1.0}
+		vec6 := []float32{-1.0, -1.0, -1.0}
 
 		similarity, err := CosineSimilarity(vec5, vec6)
 		if err != nil {
@@ -54,7 +54,7 @@ func TestCosineSimilarity(t *testing.T) {
 		}
 
 		// 方向相反的向量余弦相似度应该为 -1.0
-		if math.Abs(similarity+1.0) > 1e-10 {
+		if math.Abs(float64(similarity+1.0)) > 1e-10 {
 			t.Errorf("Expected similarity to be -1.0, got %f", similarity)
 		}
 
@@ -63,8 +63,8 @@ func TestCosineSimilarity(t *testing.T) {
 
 	// 测试用例4: 错误处理 - 向量长度不同
 	t.Run("different length vectors", func(t *testing.T) {
-		vec7 := []float64{1.0, 2.0}
-		vec8 := []float64{1.0, 2.0, 3.0}
+		vec7 := []float32{1.0, 2.0}
+		vec8 := []float32{1.0, 2.0, 3.0}
 
 		_, err := CosineSimilarity(vec7, vec8)
 		if err == nil {
@@ -76,8 +76,8 @@ func TestCosineSimilarity(t *testing.T) {
 
 	// 测试用例5: 包含零向量的情况
 	t.Run("zero vector", func(t *testing.T) {
-		vec9 := []float64{0.0, 0.0, 0.0}
-		vec10 := []float64{5.0, 6.0, 7.0}
+		vec9 := []float32{0.0, 0.0, 0.0}
+		vec10 := []float32{5.0, 6.0, 7.0}
 
 		similarity, err := CosineSimilarity(vec9, vec10)
 		if err != nil {
@@ -85,7 +85,7 @@ func TestCosineSimilarity(t *testing.T) {
 		}
 
 		// 零向量与任何向量的余弦相似度都为 0.0
-		if math.Abs(similarity) > 1e-10 {
+		if math.Abs(float64(similarity)) > 1e-10 {
 			t.Errorf("Expected similarity to be 0.0, got %f", similarity)
 		}
 
@@ -94,7 +94,7 @@ func TestCosineSimilarity(t *testing.T) {
 
 	// 附加测试：相同向量
 	t.Run("identical vectors", func(t *testing.T) {
-		vec11 := []float64{3.0, 4.0, 5.0}
+		vec11 := []float32{3.0, 4.0, 5.0}
 
 		similarity, err := CosineSimilarity(vec11, vec11)
 		if err != nil {
@@ -102,7 +102,7 @@ func TestCosineSimilarity(t *testing.T) {
 		}
 
 		// 相同向量的余弦相似度应为 1.0
-		if math.Abs(similarity-1.0) > 1e-10 {
+		if math.Abs(float64(similarity-1.0)) > 1e-10 {
 			t.Errorf("Expected similarity to be 1.0, got %f", similarity)
 		}
 
@@ -112,11 +112,11 @@ func TestCosineSimilarity(t *testing.T) {
 
 // 添加对 dotProduct 和 magnitude 函数的单元测试
 func TestDotProduct(t *testing.T) {
-	a := []float64{1, 2, 3}
-	b := []float64{4, 5, 6}
+	a := []float32{1, 2, 3}
+	b := []float32{4, 5, 6}
 
 	// 手动计算结果：1*4 + 2*5 + 3*6 = 4 + 10 + 18 = 32
-	expected := 32.0
+	expected := float32(32.0)
 
 	result, err := dotProduct(a, b)
 	if err != nil {
@@ -128,7 +128,7 @@ func TestDotProduct(t *testing.T) {
 	}
 
 	// 测试长度不同的情况
-	c := []float64{1, 2}
+	c := []float32{1, 2}
 	_, err = dotProduct(a, c)
 	if err == nil {
 		t.Errorf("Expected error for vectors with different lengths")
@@ -136,10 +136,10 @@ func TestDotProduct(t *testing.T) {
 }
 
 func TestMagnitude(t *testing.T) {
-	vec := []float64{3, 4}
+	vec := []float32{3, 4}
 
 	// 手动计算结果：sqrt(3^2 + 4^2) = sqrt(9 + 16) = sqrt(25) = 5
-	expected := 5.0
+	expected := float32(5.0)
 
 	result := magnitude(vec)
 	if result != expected {
@@ -147,7 +147,7 @@ func TestMagnitude(t *testing.T) {
 	}
 
 	// 测试零向量
-	zeroVec := []float64{0, 0, 0}
+	zeroVec := []float32{0, 0, 0}
 	result = magnitude(zeroVec)
 	if result != 0 {
 		t.Errorf("Expected magnitude of zero vector to be 0, got %f", result)
@@ -156,8 +156,8 @@ func TestMagnitude(t *testing.T) {
 
 // 基准测试，用于性能评估
 func BenchmarkCosineSimilarity(b *testing.B) {
-	vec1 := []float64{1.0, 2.0, 3.0, 4.0, 5.0}
-	vec2 := []float64{5.0, 4.0, 3.0, 2.0, 1.0}
+	vec1 := []float32{1.0, 2.0, 3.0, 4.0, 5.0}
+	vec2 := []float32{5.0, 4.0, 3.0, 2.0, 1.0}
 
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
